requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Deprecated==1.2.14
Django==4.2.6
django-extensions==3.2.3
django-oauth-plus==2.2.9
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
oauth2==1.9.0.post1
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-barcode==0.15.1
pytz==2023.3.post1
requests==2.31.0
shortuuid==1.0.11
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
wrapt==1.15.0
Pillow==10.1.0